+++ ktlint-ruleset-experimental/src/main/kotlin/com/pinterest/ktlint/ruleset/experimental/ExperimentalRuleSetProvider.kt
@@ -3,11 +3,12 @@ package com.pinterest.ktlint.ruleset.experimental
 import com.pinterest.ktlint.core.RuleSet
 import com.pinterest.ktlint.core.RuleSetProvider
 
-class ExperimentalRuleSetProvider : RuleSetProvider {
+public class ExperimentalRuleSetProvider : RuleSetProvider {
 
     override fun get(): RuleSet = RuleSet(
         "experimental",
         AnnotationRule(),
+        ArgumentListWrappingRule(),
         MultiLineIfElseRule(),
         NoEmptyFirstLineInMethodBlockRule(),
         PackageNameRule(),
@@ -16,6 +17,7 @@ class ExperimentalRuleSetProvider : RuleSetProvider {
         SpacingBetweenDeclarationsWithCommentsRule(),
         SpacingBetweenDeclarationsWithAnnotationsRule(),
         SpacingAroundAngleBracketsRule(),
-        SpacingAroundUnaryOperatorRule()
+        SpacingAroundUnaryOperatorRule(),
+        AnnotationSpacingRule()
     )
 }